+++ zipline-gradle-plugin/src/main/kotlin/app/cash/zipline/gradle/ZiplinePlugin.kt
@@ -58,12 +58,23 @@ class ZiplinePlugin : KotlinCompilerPluginSupportPlugin {
     val ziplineExtension = target.extensions.create("zipline", ZiplineExtension::class.java)
     ziplineExtension.apiTracking.convention(true)
 
+    val cliConfiguration: Configuration = target.configurations.create("ziplineCli")
+      .apply {
+        isCanBeConsumed = false
+        isVisible = false
+      }
+    target.dependencies.add(
+      cliConfiguration.name,
+      target.ziplineDependency("zipline-cli"),
+    )
+
     kotlinExtension.targets.withType(KotlinJsIrTarget::class.java).all { kotlinTarget ->
       kotlinTarget.binaries.withType(JsIrBinary::class.java).all { kotlinBinary ->
         registerCompileZiplineTask(
           project = target,
           jsProductionTask = kotlinBinary.asJsProductionTask(),
           extension = ziplineExtension,
+          cliConfiguration = cliConfiguration,
         )
       }
     }
@@ -77,6 +88,7 @@ class ZiplinePlugin : KotlinCompilerPluginSupportPlugin {
         project = target,
         jsProductionTask = jsProductionTask,
         extension = ziplineExtension,
+        cliConfiguration = cliConfiguration,
       )
       ziplineCompileTask.configure {
         it.dependsOn(kotlinWebpack)
@@ -88,16 +100,6 @@ class ZiplinePlugin : KotlinCompilerPluginSupportPlugin {
       )
       kotlinWebpack.dependsOn(writeWebpackConfigTask)
     }
-
-    val cliConfiguration: Configuration = target.configurations.create("ziplineCli")
-      .apply {
-        isCanBeConsumed = false
-        isVisible = false
-      }
-    target.dependencies.add(
-      cliConfiguration.name,
-      target.ziplineDependency("zipline-cli"),
-    )
 
     target.afterEvaluate {
       if (ziplineExtension.apiTracking.get()) {
@@ -121,6 +123,7 @@ class ZiplinePlugin : KotlinCompilerPluginSupportPlugin {
     project: Project,
     jsProductionTask: JsProductionTask,
     extension: ZiplineExtension,
+    cliConfiguration: Configuration,
   ): TaskProvider<ZiplineCompileTask> {
     val target = (if (jsProductionTask.targetName == "js") "" else jsProductionTask.targetName)
     val mode = jsProductionTask.mode.name
@@ -133,9 +136,8 @@ class ZiplinePlugin : KotlinCompilerPluginSupportPlugin {
     val ziplineCompileTask = project.tasks.register(
       "${jsProductionTask.name}Zipline",
       ZiplineCompileTask::class.java,
-    )
-    ziplineCompileTask.configure {
-      it.configure(outputDirectoryName, jsProductionTask, extension)
+    ) {
+      it.configure(outputDirectoryName, jsProductionTask, extension, cliConfiguration)
     }
 
     val serveTaskName = "serve${target.capitalize()}${mode.capitalize()}${toolName}Zipline"